command-service/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.1.8
    # via flask
flask==3.1.2
    # via -r requirements.in
idna==3.11
    # via requests
importlib-metadata==8.7.0
    # via flask
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
pika==1.3.2
    # via -r requirements.in
psycopg2-binary==2.9.11
    # via -r requirements.in
python-dotenv==1.2.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via flask
zipp==3.23.0
    # via importlib-metadata
